+++ packages/cloudflare/src/wfp-provisioner/tenant-hook.ts
@@ -1,5 +1,6 @@
 import type { TenantsDataAdapter } from "@authhero/adapter-interfaces";
 import type { CloudflareWfpD1Provisioner } from "./types";
+import { collectSyncDefaultsErrors } from "./sync-defaults-errors";
 
 /**
  * Adapt the provisioner to `@authhero/multi-tenancy`'s
@@ -119,31 +120,6 @@ export interface WfpTenantProvisioningHook {
 
 function defaultShouldProvision(tenant: { deployment_type?: string }): boolean {
   return tenant.deployment_type === "wfp";
-}
-
-/**
- * Collects per-entity `errors` from a sync-defaults apply result. The seed runs
- * with `continueOnError`, so the tenant worker returns a 2xx (no rejection)
- * even when individual entities fail — a clean resolve is therefore not proof
- * the seed landed. Walk the result's entity outcomes and surface any collected
- * errors so a partially-seeded tenant isn't marked `ready`.
- */
-function collectSyncDefaultsErrors(result: unknown): string[] {
-  if (typeof result !== "object" || result === null) return [];
-  const errors: string[] = [];
-  for (const outcome of Object.values(result)) {
-    if (
-      typeof outcome === "object" &&
-      outcome !== null &&
-      "errors" in outcome &&
-      Array.isArray(outcome.errors)
-    ) {
-      for (const err of outcome.errors) {
-        if (typeof err === "string") errors.push(err);
-      }
-    }
-  }
-  return errors;
 }
 
 // Cap persisted error text (tenant rows and reported step details) so a
